Refactor test harness to support test pragmas
